requirements.txt
matplotlib
numpy
pynetworktables
